XWIKI-12788: Introduce skin extension plugins for webjar resources
