+++ xwiki-platform-core/xwiki-platform-webjars/xwiki-platform-webjars-api/src/test/java/org/xwiki/webjars/internal/WebjarUrlResolverTest.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.xwiki.webjars;
+package org.xwiki.webjars.internal;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -47,13 +47,12 @@ import static org.mockito.Mockito.when;
  * Unit tests for {@link org.xwiki.webjars.script.WebJarsScriptService}.
  *
  * @version $Id$
- * @since 6.0M1
  */
 @ComponentTest
-class WebJarsScriptServiceTest
+class WebjarUrlResolverTest
 {
     @InjectMockComponents
-    private WebJarsScriptService scriptService;
+    private WebjarUrlResolver scriptService;
 
     @MockComponent
     private ResourceReferenceSerializer<ResourceReference, ExtendedURL> serializer;
